half4 -> float4 in frag prog

+++ Shaders/Resources/UI-Default-SoftMaskable.shader
@@ -99,7 +99,7 @@ Shader "Hidden/UI/Default (SoftMaskable)"
 
             fixed4 frag(v2f IN) : SV_Target
             {
-                half4 color = (tex2D(_MainTex, IN.texcoord) + _TextureSampleAdd) * IN.color;
+                float4 color = (tex2D(_MainTex, IN.texcoord) + _TextureSampleAdd) * IN.color;
 
                 color.a *= UnityGet2DClipping(IN.worldPosition.xy, _ClipRect);
